# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "my project doc",
        "description": "\u6e38\u620f\u6570\u636e\u8bc4\u4f30\u63a5\u53e3\u6587\u6863, V1.0<br>",
        "version": "1.0"
    },
    "host": "assess.act.sdo.com:81",
    "schemes": [
        "http"
    ],
    "consumes": [
        "multipart/form-data"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/index.php/UserScaleModel/evaluate": {
            "get": {
                "tags": [
                    "UserScaleModel"
                ],
                "summary": "\u9884\u4f30\u7528\u6237\u6a21\u578b",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "\u6e38\u620f\u540d\u79f0",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "\u6e38\u620f\u7c7b\u578b",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "\u6d4b\u8bd5\u6027\u8d28",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "ip",
                        "in": "query",
                        "description": "\u662f\u5426\u6709ip",
                        "required": false,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "scale",
                        "in": "query",
                        "description": "\u7528\u6237\u89c4\u6a21",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "1dayret",
                        "in": "query",
                        "description": "\u6b21\u7559",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "3dayret",
                        "in": "query",
                        "description": "3\u7559",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "7dayret",
                        "in": "query",
                        "description": "7\u7559",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "30dayret",
                        "in": "query",
                        "description": "30\u65e5\u7559",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "1dayltv",
                        "in": "query",
                        "description": "\u4e00\u5929LTV",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "7dayltv",
                        "in": "query",
                        "description": "7\u5929LTV",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "30dayltv",
                        "in": "query",
                        "description": "30\u5929LTV",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "modified_scale",
                        "in": "query",
                        "description": "\u4fee\u6539\u540e\u7684\u7528\u6237\u89c4\u6a21",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "\u64cd\u4f5c\u6210\u529f",
                        "schema": {
                            "properties": {
                                "UserScale": {
                                    "description": "\u7528\u6237\u89c4\u6a21",
                                    "type": "integer"
                                },
                                "first_day_ret": {
                                    "description": "\u6b21\u7559",
                                    "type": "number"
                                },
                                "third_day_ret": {
                                    "description": "3\u7559",
                                    "type": "number"
                                },
                                "seven_day_ret": {
                                    "description": "7\u7559",
                                    "type": "number"
                                },
                                "thirty_day_ret": {
                                    "description": "30\u7559",
                                    "type": "number"
                                },
                                "first_day_ltv": {
                                    "description": "1\u5929LTV",
                                    "type": "number"
                                },
                                "seven_day_ltv": {
                                    "description": "7\u5929LTV",
                                    "type": "number"
                                },
                                "thirty_day_ltv": {
                                    "description": "30\u5929LTV",
                                    "type": "number"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}